# backend/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a API for cakrawala.",
        "title": "Cakrawala API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Kelompok 14"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/admin/add-reward": {
            "post": {
                "description": "Add New Reward.",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authentication"
                ],
                "summary": "AddReward.",
                "parameters": [
                    {
                        "description": "Inputan yang benar",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/transactions.AddRewardBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/v1/exchange-reward": {
            "post": {
                "description": "Exchange Reward with Points.",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authentication"
                ],
                "summary": "ExchangeReward.",
                "parameters": [
                    {
                        "description": "Inputan yang benar",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/transactions.ExchangeRewardBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "transactions.AddRewardBody": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "stock": {
                    "type": "integer"
                }
            }
        },
...
        "transactions.ExchangeRewardBody": {
            "type": "object",
            "properties": {
                "qty": {
                    "type": "integer"
                },
                "reward_id": {
                    "type": "integer"
                }
            }
        },
...
    }
}